moved auth logic to custom hook
